+++ LMWNCoin/Scenes/Dashboard/Invitefriend/InvitefriendTableViewCell.swift
@@ -1,0 +1,6 @@
+
+import UIKit
+
+class InvitefriendTableViewCell: UITableViewCell {
+
+}
