--- LMWNCoin/Scenes/Dashboard/InvitefriendTableViewCell.swift
@@ -1,12 +1,0 @@
-//
-//  InvitefriendTableViewCell.swift
-//  LMWNCoin
-//
-//  Created by kowit nanumchill on 31/3/2567 BE.
-//
-
-import UIKit
-
-class InvitefriendTableViewCell: UITableViewCell {
-
-}
