+++ LMWNCoin/Scenes/Dashboard/ListCoins/DashboardTableViewCell.swift
@@ -1,9 +1,3 @@
-//
-//  DashboardTableViewCell.swift
-//  LMWNCoin
-//
-//  Created by kowit nanumchill on 28/3/2567 BE.
-//
 
 import UIKit
 import SDWebImage
